Series of patches with minor matching optimizations.

ab65905 - adds BenchmarkQueryRange to measure effect of future changes
 e222d40 - avoids extra iteration over possible results, run matchers checks before adding to a result instead 
 1121364 - extracts logic which prepares candidate FPs into a separate method
9b15a3b - adds `fpsForLabelMatchers` and `seriesForLabelMatchers` methods and uses them where appropriate 

Before (as of ab65905):
```
go test -bench BenchmarkQueryRange -run none -benchtime=10s -benchmem
    1000	  14723524 ns/op	 5013534 B/op	   83059 allocs/op
```

After:
```
go test -bench BenchmarkQueryRange -run none -benchtime=10s -benchmem
    1000	  11553489 ns/op	 4688423 B/op	   82596 allocs/op
```