@redbaron 2016-09-19
rebased