@redbaron 2016-09-20
Can someone have a look please? It is up to 30% speedup on QueryRange benchmark