@redbaron 2016-10-02
@beorn7 , made suggested changes