@redbaron 2016-10-04
sorry, I didn't  know this PR is holding a release back.  I'll keep this tab open and expect to be able to make fixes if needed within an hour.