@beorn7 2016-10-05
Thanks. The only missing nit is the one about the loop labels. See comments above. After that, we can merge.