@beorn7 2016-10-05
Sorry for the nit, but ALLCAPS is also not very Go idiomatic. Please use `FPLoop` as label name.