@beorn7 2016-10-31
In principle, yes. It definitely needs a major update. I had a script for it somewhere…